refactored entire defs_provider into macrostrat.core
