+++ py-modules/column-ingestion/macrostrat/column_ingestion/project_metadata/test_project_metadata.py
@@ -17,7 +17,7 @@ from . import _column_metadata_importer
 
 __here__ = Path(__file__).parent
 
-from ..defs_provider import MacrostratDatabaseDataProvider, MacrostratMetadataPopulator
+from macrostrat.core.defs_provider import MacrostratDatabaseDataProvider, MacrostratMetadataPopulator
 
 
 @fixture(scope="class")
