+++ py-modules/core/macrostrat/core/defs_provider/sql/load-intervals.sql
